refactored models.py already
